requirements.txt
autogen-ext[openai,azure]
azure-identity
python-dotenv
autogen-agentchat
autogen-core
pandas
langchain>=0.0.225
langchain-experimental
